rename tests be recognized by maven
