--- qoffer-domain/src/test/groovy/life/qbic/portal/portlet/offers/OfferTest.groovy
@@ -1,14 +1,0 @@
-package life.qbic.portal.portlet.offers
-
-import spock.lang.Specification
-
-/**
- * <short description>
- *
- * <detailed description>
- *
- * @since: 1.0.0
- * @author: Tobias Koch
- */
-class OfferTest extends Specification {
-}
